切换到Hibernate

+++ src/main/kotlin/com/enihsyou/astolfo/hotel/domain/Room.kt
@@ -1,7 +1,5 @@
 package com.enihsyou.astolfo.hotel.domain
 
-import java.math.BigInteger
-
 enum class RoomType {
   大床房, 单人间, 双人间, 总统套房
 }
@@ -10,8 +8,11 @@ enum class RoomDirection {
   东, 西, 南, 北
 }
 
+
+
 data class Room(
-    val id: BigInteger,
+
+    val id: Long,
     val type: RoomType,
     val floor: Int,
     val direction: RoomDirection,
